requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
jinja2==3.1.6
python-multipart==0.0.20
passlib==1.7.4
pydantic-settings==2.10.1
httpx==0.28.1
openai==1.101.0
google-genai==1.30.0
pillow==11.3.0
python-dotenv==1.1.1
itsdangerous==2.2.0
pytest==8.4.1
